+++ scenario-compiler/src/test/scala/pl/touk/nussknacker/engine/compile/ProcessValidatorSpec.scala
@@ -417,7 +417,7 @@ class ProcessValidatorSpec extends AnyFunSuite with Matchers with Inside with Op
       ListMap(
         "Field1" -> Typed.fromInstance("Field1Value"),
         "Field2" -> Typed.fromInstance("Field2Value"),
-        "Field3" -> Typed[BigDecimal]
+        "Field3" -> Typed[java.math.BigDecimal]
       )
     )
     varsType.fields.get("spelVariable").value shouldEqual Typed[Boolean]
@@ -1747,8 +1747,8 @@ class ProcessValidatorSpec extends AnyFunSuite with Matchers with Inside with Op
 
   case class SimpleRecord(
       value1: AnotherSimpleRecord,
-      plainValue: BigDecimal,
-      plainValueOpt: Option[BigDecimal],
+      plainValue: java.math.BigDecimal,
+      plainValueOpt: Option[java.math.BigDecimal],
       intAsAny: Any,
       list: java.util.List[SimpleRecord]
   ) {
@@ -1762,8 +1762,18 @@ class ProcessValidatorSpec extends AnyFunSuite with Matchers with Inside with Op
   case class AnotherSimpleRecord(value2: Long)
 
   class SampleEnricher extends Service {
+
     def invoke()(implicit ec: ExecutionContext): Future[SimpleRecord] =
-      Future.successful(SimpleRecord(AnotherSimpleRecord(1), 2, Option(2), 1, Collections.emptyList[SimpleRecord]))
+      Future.successful(
+        SimpleRecord(
+          AnotherSimpleRecord(1),
+          new java.math.BigDecimal(1),
+          Option(new java.math.BigDecimal(2)),
+          1,
+          Collections.emptyList[SimpleRecord]
+        )
+      )
+
   }
 
   object ProcessHelper {